created Drawer component
